+++ .gitignore
@@ -15,6 +15,7 @@
 # production
 /build
 /.yarn
+/.yarn/install-state.gz
 
 # misc
 .DS_Store
