+++ src/components/BaseNavi.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useState } from "react";
 import NextImage from "next/image";
 import MenuBox from "./MenuBox";
@@ -43,7 +44,9 @@ export default function BaseNavi() {
           </nav>
         </div>
       </header>
-      {_switch && <MenuBox />}
+      {_switch && <MenuBox menuAction={() => { 
+        _setSwitch(!_switch);
+       }} />}
     </>
   );
 }
