+++ src/styles/parts/_contents.sass
@@ -100,7 +100,7 @@
 
 .about-section
   position: relative
-  top: 38px
+  top: 48px
   max-height: calc(100vh - 48px)
   min-height: calc(100vh - 48px)
   max-width: 50%
